made partners component
fixe the whole folder structure,deleted unnecesssary files and created the partners component
