+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Navbar from "./Components/Navbar";
+import Navbar from"@/components/Navbar";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
